06/02/2023

+++ modelo/personalModelo.php
@@ -1,20 +1,12 @@
-<?php 
+<?php
 require_once "conexion.php";
-class ModeloPersonal{
+class ModeloPersonal
+{
 
 
-  static public function mdlListaPersonal(){
-    $stmt=Conexion::conectar()->prepare("select * from personal");
-    $stmt->execute();
-
-    return $stmt->fetchAll();
-
-    $stmt->close();
-    $stmt->null;
-  }
-  
-   static public function mdlInfoChoferes(){
-    $stmt=Conexion::conectar()->prepare("select * from personal where cargo_personal='Chofer' and estado_personal=1");
+  static public function mdlListaPersonal()
+  {
+    $stmt = Conexion::conectar()->prepare("select * from personal");
     $stmt->execute();
 
     return $stmt->fetchAll();
@@ -23,8 +15,20 @@ class ModeloPersonal{
     $stmt->null;
   }
 
-  static public function mdlInfoVentasPersonal(){
-    $stmt=Conexion::conectar()->prepare("select * from personal where cargo_personal='Chofer' or cargo_personal='Secretaria' ");
+  static public function mdlInfoChoferes()
+  {
+    $stmt = Conexion::conectar()->prepare("select * from personal where cargo_personal='Chofer' and estado_personal=1");
+    $stmt->execute();
+
+    return $stmt->fetchAll();
+
+    $stmt->close();
+    $stmt->null;
+  }
+
+  static public function mdlInfoVentasPersonal()
+  {
+    $stmt = Conexion::conectar()->prepare("select * from personal where cargo_personal='Chofer' or cargo_personal='Secretaria' ");
     $stmt->execute();
     return $stmt->fetchAll();
 
@@ -32,8 +36,9 @@ class ModeloPersonal{
     $stmt->null;
   }
 
-  static public function mdlInfoPersonalDisp(){
-    $stmt=Conexion::conectar()->prepare("select * from personal where cargo_personal='Chofer' and estado_personal=1 or cargo_personal='Secretaria' and estado_personal=1");
+  static public function mdlInfoPersonalDisp()
+  {
+    $stmt = Conexion::conectar()->prepare("select * from personal where cargo_personal='Chofer' and estado_personal=1 or cargo_personal='Secretaria' and estado_personal=1");
     $stmt->execute();
     return $stmt->fetchAll();
 
@@ -41,24 +46,25 @@ class ModeloPersonal{
     $stmt->null;
   }
 
-  static public function mdlRegPersonal($data){
+  static public function mdlRegPersonal($data)
+  {
 
-    $nomPersonal=$data["nomPersonal"];
-    $patPersonal=$data["patPersonal"];
-    $matPersonal=$data["matPersonal"];
-    $ciPersonal=$data["ciPersonal"];
-    $telPersonal=$data["telPersonal"];
-    $contactoReferencia=$data["contactoReferencia"];
-    $cargoPersonal=$data["cargoPersonal"];
-    $fechaNacPersonal=$data["fechaNacPersonal"];
-    $fechaIngPersonal=$data["fechaIngPersonal"];
+    $nomPersonal = $data["nomPersonal"];
+    $patPersonal = $data["patPersonal"];
+    $matPersonal = $data["matPersonal"];
+    $ciPersonal = $data["ciPersonal"];
+    $telPersonal = $data["telPersonal"];
+    $contactoReferencia = $data["contactoReferencia"];
+    $cargoPersonal = $data["cargoPersonal"];
+    $fechaNacPersonal = $data["fechaNacPersonal"];
+    $fechaIngPersonal = $data["fechaIngPersonal"];
 
     /* var_dump($nomPersonal); */
-    $stmt=Conexion::conectar()->prepare("insert into personal(nombre_personal, ap_pat_personal, ap_mat_personal, ci_personal, cargo_personal, fecha_nac_personal, telefono_personal, contacto_referencia, fecha_ingreso) values('$nomPersonal', '$patPersonal', '$matPersonal', '$ciPersonal', '$cargoPersonal', '$fechaNacPersonal', '$telPersonal', '$contactoReferencia', '$fechaIngPersonal')");
+    $stmt = Conexion::conectar()->prepare("insert into personal(nombre_personal, ap_pat_personal, ap_mat_personal, ci_personal, cargo_personal, fecha_nac_personal, telefono_personal, contacto_referencia, fecha_ingreso) values('$nomPersonal', '$patPersonal', '$matPersonal', '$ciPersonal', '$cargoPersonal', '$fechaNacPersonal', '$telPersonal', '$contactoReferencia', '$fechaIngPersonal')");
 
-    if($stmt->execute()){
+    if ($stmt->execute()) {
       return "ok";
-    }else{
+    } else {
       return "error";
     }
 
@@ -66,8 +72,9 @@ class ModeloPersonal{
     $stmt->null;
   }
 
-  static public function mdlInfoPersonal($id){
-    $stmt=Conexion::conectar()->prepare("select * from personal where id_personal=$id");
+  static public function mdlInfoPersonal($id)
+  {
+    $stmt = Conexion::conectar()->prepare("select * from personal where id_personal=$id");
     $stmt->execute();
 
     return $stmt->fetch();
@@ -76,25 +83,26 @@ class ModeloPersonal{
     $stmt->null;
   }
 
-  static public function mdlEditPersonal($data){
+  static public function mdlEditPersonal($data)
+  {
 
-    $idPersonal=$data["idPersonal"];
-    $nomPersonal=$data["nomPersonal"];
-    $patPersonal=$data["patPersonal"];
-    $matPersonal=$data["matPersonal"];
-    $ciPersonal=$data["ciPersonal"];
-    $telPersonal=$data["telPersonal"];
-    $contactoReferencia=$data["contactoReferencia"];
-    $cargoPersonal=$data["cargoPersonal"];
-    $fechaNacPersonal=$data["fechaNacPersonal"];
-    $fechaIngPersonal=$data["fechaIngPersonal"];
-    $estadoPersonal=$data["estadoPersonal"];
+    $idPersonal = $data["idPersonal"];
+    $nomPersonal = $data["nomPersonal"];
+    $patPersonal = $data["patPersonal"];
+    $matPersonal = $data["matPersonal"];
+    $ciPersonal = $data["ciPersonal"];
+    $telPersonal = $data["telPersonal"];
+    $contactoReferencia = $data["contactoReferencia"];
+    $cargoPersonal = $data["cargoPersonal"];
+    $fechaNacPersonal = $data["fechaNacPersonal"];
+    $fechaIngPersonal = $data["fechaIngPersonal"];
+    $estadoPersonal = $data["estadoPersonal"];
 
-    $stmt=Conexion::conectar()->prepare("update personal set nombre_personal='$nomPersonal', ap_pat_personal='$patPersonal', ap_mat_personal='$matPersonal', ci_personal='$ciPersonal', cargo_personal='$cargoPersonal', fecha_nac_personal='$fechaNacPersonal', telefono_personal='$telPersonal', contacto_referencia='$contactoReferencia', fecha_ingreso='$fechaIngPersonal', estado_personal='$estadoPersonal' where id_personal=$idPersonal");
+    $stmt = Conexion::conectar()->prepare("update personal set nombre_personal='$nomPersonal', ap_pat_personal='$patPersonal', ap_mat_personal='$matPersonal', ci_personal='$ciPersonal', cargo_personal='$cargoPersonal', fecha_nac_personal='$fechaNacPersonal', telefono_personal='$telPersonal', contacto_referencia='$contactoReferencia', fecha_ingreso='$fechaIngPersonal', estado_personal='$estadoPersonal' where id_personal=$idPersonal");
 
-    if($stmt->execute()){
+    if ($stmt->execute()) {
       return "ok";
-    }else{
+    } else {
       return "error";
     }
 
@@ -102,19 +110,25 @@ class ModeloPersonal{
     $stmt->null;
   }
 
-  static public function mdlEliPersonal($data){
-    
-    $personal=Conexion::conectar()->prepare("select * from nota_entrega where id_personal=$data");
-    $personal->execute();
-    if($personal->fetch()>0){
-      echo "error";
-    }else{
-      $stmt=Conexion::conectar()->prepare("delete from personal where id_personal=$data");
+  static public function mdlEliPersonal($data)
+  {
 
-      if($stmt->execute()){
-        return "ok";
-      }else{
-        return "error";
+    $personal = Conexion::conectar()->prepare("select * from nota_entrega where id_personal=$data");
+    $personal->execute();
+    if ($personal->fetch() > 0) {
+      echo "error";
+    } else {
+      $factura = Conexion::conectar()->prepare("select * from factura where id_personal=$data");
+      $factura->execute();
+      if ($factura->fetch() > 0) {
+        echo "error";
+      } else {
+        $stmt = Conexion::conectar()->prepare("delete from personal where id_personal=$data");
+        if ($stmt->execute()) {
+          return "ok";
+        } else {
+          return "error";
+        }
       }
     }
 
@@ -122,18 +136,9 @@ class ModeloPersonal{
     $stmt->null;
   }
 
-  static public function mdlBusPersonal($data){
-    $stmt=Conexion::conectar()->prepare("select * from Personal where nit_ci_Personal=$data");
-    $stmt->execute();
-
-    return $stmt->fetch();
-
-    $stmt->close();
-    $stmt->null;
-  }
-  
-  static public function mdlCantidadPersonals(){
-    $stmt=Conexion::conectar()->prepare("select count(*) as personal from personal");
+  static public function mdlBusPersonal($data)
+  {
+    $stmt = Conexion::conectar()->prepare("select * from Personal where nit_ci_Personal=$data");
     $stmt->execute();
 
     return $stmt->fetch();
@@ -142,4 +147,14 @@ class ModeloPersonal{
     $stmt->null;
   }
 
+  static public function mdlCantidadPersonals()
+  {
+    $stmt = Conexion::conectar()->prepare("select count(*) as personal from personal");
+    $stmt->execute();
+
+    return $stmt->fetch();
+
+    $stmt->close();
+    $stmt->null;
+  }
 }
